+++ lib/core/widgets/bottom_navigation_bar/center_action_button.dart
@@ -1,0 +1,58 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:routina/core/theaming/app_colors.dart';
+
+class CenterActionButton extends StatelessWidget {
+  final VoidCallback onTap;
+  final VoidCallback? onLongPress;
+
+  const CenterActionButton({
+    super.key,
+    required this.onTap,
+    this.onLongPress,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Expanded(
+      child: GestureDetector(
+        onTap: () {
+          HapticFeedback.mediumImpact();
+          onTap();
+        },
+        onLongPress: () {
+          if (onLongPress != null) {
+            HapticFeedback.heavyImpact();
+            onLongPress!();
+          }
+        },
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Container(
+              width: 46.w,
+              height: 46.w,
+              decoration: BoxDecoration(
+                shape: BoxShape.circle,
+                gradient: const LinearGradient(
+                  colors: [AppColors.primary, AppColors.primaryDark],
+                  begin: Alignment.topLeft,
+                  end: Alignment.bottomRight,
+                ),
+                boxShadow: [
+                  BoxShadow(
+                    color: AppColors.primary.withValues(alpha: 0.4),
+                    blurRadius: 12,
+                    offset: const Offset(0, 4),
+                  ),
+                ],
+              ),
+              child: Icon(Icons.add_rounded, color: Colors.white, size: 28.sp),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
