+++ lib/core/widgets/main_navigation_bar.dart
@@ -1,19 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:routina/core/helpers/spacing.dart';
 import 'package:routina/core/theaming/app_colors.dart';
 import 'package:routina/core/widgets/create_habit/ui/create_habit_bottom_sheet.dart';
 import 'package:routina/core/widgets/logout_button/cubit/logout_cubit.dart';
 import 'package:routina/features/analyze_screen/logic/cubit/ai_analysis_cubit.dart';
 import 'package:routina/features/analyze_screen/ui/widgets/quick_ai_analysis_sheet.dart';
 import 'package:routina/features/home_screen/logic/cubit/home_cubit.dart';
-import 'package:routina/features/home_screen/ui/home_screen.dart';
-import 'package:routina/features/habit_tracker_screen/ui/habit_tracker_screen.dart';
-import 'package:routina/features/analyze_screen/ui/analyze_screen.dart';
 import 'package:routina/features/profile_screen/logic/cubit/profile_cubit.dart';
-import 'package:routina/features/profile_screen/ui/profile_screen.dart';
+import 'package:routina/core/widgets/bottom_navigation_bar/screen_selector.dart';
+import 'package:routina/core/widgets/bottom_navigation_bar/bottom_navigation_bar_widget.dart';
 
 class MainNavigationBar extends StatefulWidget {
   const MainNavigationBar({super.key});
@@ -25,15 +20,8 @@ class MainNavigationBar extends StatefulWidget {
 class _MainNavigationBarState extends State<MainNavigationBar> {
   int _currentIndex = 0;
 
-  int _getScreenIndex(int navIndex) {
-    if (navIndex < 2) return navIndex;
-    if (navIndex > 2) return navIndex - 1;
-    return 0;
-  }
-
   @override
   Widget build(BuildContext context) {
-    final isDark = Theme.of(context).brightness == Brightness.dark;
 
     return MultiBlocProvider(
       providers: [
@@ -45,199 +33,41 @@ class _MainNavigationBarState extends State<MainNavigationBar> {
       child: Builder(
         builder: (context) {
           return Scaffold(
-            body: _buildCurrentScreen(),
-            bottomNavigationBar: _buildBottomNav(context, isDark),
-          );
-        },
-      ),
-    );
-  }
-
-  Widget _buildCurrentScreen() {
-    switch (_getScreenIndex(_currentIndex)) {
-      case 0:
-        return const HomeScreen();
-      case 1:
-        return const HabitTrackerScreen();
-      case 2:
-        return const AnalyzeScreen();
-      case 3:
-        return const ProfileScreen();
-      default:
-        return const HomeScreen();
-    }
-  }
-
-  Widget _buildBottomNav(BuildContext context, bool isDark) {
-    final bgColor = isDark ? AppColors.darkSurface : AppColors.surface;
-    final selectedColor = AppColors.primary;
-    final unselectedColor = isDark
-        ? AppColors.darkTextSecondary
-        : AppColors.textSecondary;
-
-    return Container(
-      decoration: BoxDecoration(
-        color: bgColor,
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withValues(alpha:  isDark ? 0.3 : 0.08),
-            blurRadius: 20,
-            offset: const Offset(0, -4),
-          ),
-        ],
-      ),
-      child: SafeArea(
-        child: SizedBox(
-          height: 64.h,
-          child: Row(
-            children: [
-              _buildNavItem(
-                context,
-                index: 0,
-                icon: Icons.home_outlined,
-                activeIcon: Icons.home_rounded,
-                label: 'Home',
-                selectedColor: selectedColor,
-                unselectedColor: unselectedColor,
-              ),
-              _buildNavItem(
-                context,
-                index: 1,
-                icon: Icons.track_changes_outlined,
-                activeIcon: Icons.track_changes_rounded,
-                label: 'Habits',
-                selectedColor: selectedColor,
-                unselectedColor: unselectedColor,
-              ),
-              _buildCenterButton(context),
-              _buildNavItem(
-                context,
-                index: 3,
-                icon: Icons.analytics_outlined,
-                activeIcon: Icons.analytics_rounded,
-                label: 'Analyze',
-                selectedColor: selectedColor,
-                unselectedColor: unselectedColor,
-              ),
-              _buildNavItem(
-                context,
-                index: 4,
-                icon: Icons.person_outline_rounded,
-                activeIcon: Icons.person_rounded,
-                label: 'Profile',
-                selectedColor: selectedColor,
-                unselectedColor: unselectedColor,
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildNavItem(
-    BuildContext context, {
-    required int index,
-    required IconData icon,
-    required IconData activeIcon,
-    required String label,
-    required Color selectedColor,
-    required Color unselectedColor,
-  }) {
-    final isSelected = _currentIndex == index;
-
-    return Expanded(
-      child: GestureDetector(
-        onTap: () {
-          HapticFeedback.selectionClick();
-          setState(() => _currentIndex = index);
-        },
-        behavior: HitTestBehavior.opaque,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            AnimatedSwitcher(
-              duration: const Duration(milliseconds: 200),
-              child: Icon(
-                isSelected ? activeIcon : icon,
-                key: ValueKey(isSelected),
-                color: isSelected ? selectedColor : unselectedColor,
-                size: 24.sp,
-              ),
-            ),
-            verticalSpace(4), 
-            Text(
-              label,
-              style: TextStyle(
-                fontSize: 10.sp,
-                fontWeight: isSelected ? FontWeight.w600 : FontWeight.normal,
-                color: isSelected ? selectedColor : unselectedColor,
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildCenterButton(BuildContext context) {
-    return Expanded(
-      child: GestureDetector(
-        onTap: () {
-          HapticFeedback.mediumImpact();
-          final homeCubit = context.read<HomeCubit>();
-          showModalBottomSheet(
-            context: context,
-            isScrollControlled: true,
-            backgroundColor: Colors.transparent,
-            builder: (context) => BlocProvider.value(
-              value: homeCubit,
-              child: const CreateHabitBottomSheet(),
-            ),
-          );
-        },
-        onLongPress: () {
-          final habits = context.read<HomeCubit>().state.habits;
-          if (habits.isEmpty) {
-            ScaffoldMessenger.of(context).showSnackBar(
-              const SnackBar(
-                content: Text('Add some habits first!'),
-                backgroundColor: AppColors.primary,
-                behavior: SnackBarBehavior.floating,
-              ),
-            );
-            return;
-          }
-          HapticFeedback.heavyImpact();
-          final aiCubit = context.read<AiAnalysisCubit>();
-          aiCubit.analyzeHabits(habits);
-          showQuickAiAnalysisSheet(context);
-        },
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Container(
-              width: 46.w,
-              height: 46.w,
-              decoration: BoxDecoration(
-                shape: BoxShape.circle,
-                gradient: const LinearGradient(
-                  colors: [AppColors.primary, AppColors.primaryDark],
-                  begin: Alignment.topLeft,
-                  end: Alignment.bottomRight,
-                ),
-                boxShadow: [
-                  BoxShadow(
-                    color: AppColors.primary.withValues(alpha:0.4),
-                    blurRadius: 12,
-                    offset: const Offset(0, 4),
+            body: ScreenSelector(currentIndex: _currentIndex),
+            bottomNavigationBar: BottomNavigationBarWidget(
+              currentIndex: _currentIndex,
+              onIndexChanged: (index) => setState(() => _currentIndex = index),
+              onCenterTap: () {
+                final homeCubit = context.read<HomeCubit>();
+                showModalBottomSheet(
+                  context: context,
+                  isScrollControlled: true,
+                  backgroundColor: Colors.transparent,
+                  builder: (context) => BlocProvider.value(
+                    value: homeCubit,
+                    child: const CreateHabitBottomSheet(),
                   ),
-                ],
-              ),
-              child: Icon(Icons.add_rounded, color: Colors.white, size: 28.sp),
+                );
+              },
+              onCenterLongPress: () {
+                final habits = context.read<HomeCubit>().state.habits;
+                if (habits.isEmpty) {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    const SnackBar(
+                      content: Text('Add some habits first!'),
+                      backgroundColor: AppColors.primary,
+                      behavior: SnackBarBehavior.floating,
+                    ),
+                  );
+                  return;
+                }
+                final aiCubit = context.read<AiAnalysisCubit>();
+                aiCubit.analyzeHabits(habits);
+                showQuickAiAnalysisSheet(context);
+              },
             ),
-          ],
-        ),
+          );
+        },
       ),
     );
   }
